refactor: split parser tests into domain-specific files
- parser_test.go: shared helper, Parser struct, helper function tests
- parse_mission_test.go: ParseMission tests
- parse_soldier_test.go: ParseSoldierState tests
- parse_vehicle_test.go: ParseVehicleState tests
- parse_events_test.go: ParseKillEvent tests
